WIP Attempt to split ggez up into multiple crates
