+++ crates/ggez_2d/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "ggez_2d"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+mint.workspace = true
+glam.workspace = true
+wgpu.workspace = true
+crevice.workspace = true
+glyph_brush.workspace = true
+ggez_core = { path = "../ggez_core" }
+ggez_error = { path = "../ggez_error" }
+
+[lib]
+name = "ggez_2d"
+path = "src/lib.rs"
+
