+++ crates/ggez_2d/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod canvas;
+pub mod internal_canvas;
