+++ crates/ggez_3d/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "ggez_3d"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
