+++ crates/ggez_3d/README.md
@@ -1,0 +1,1 @@
+# This crate is a placeholder and will soon be populated by an actual 3d system
