+++ crates/ggez_core/src/filesystem.rs
@@ -35,8 +35,10 @@
 use crate::{
     conf,
     vfs::{self, OverlayFS, VFS},
-    Context, GameError, GameResult,
+    Context,
 };
+use ggez_error::prelude::*;
+
 use directories::ProjectDirs;
 use std::{
     env, io,
