+++ crates/ggez_core/src/timer.rs
