+++ crates/ggez_error/src/error.rs
