+++ crates/ggez_error/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod error;
+
+pub mod prelude {
+    pub use crate::error::{GameError, GameResult};
+}
