+++ crates/ggez_graphics/src/gpu/growing.rs
