+++ crates/ggez_traits/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ggez_traits"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[lib]
+name = "ggez_traits"
+path = "src/lib.rs"
